+++ effects-cats-2/shared/src/main/scala/busymachines/pureharm/capabilities/RandomImpl.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 BusyMachines
+ * Copyright 2020-2021 Typelevel
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,45 +19,113 @@ package busymachines.pureharm.capabilities
 import java.util.UUID
 import scala.util.{Random => ScalaRandom}
 
-import busymachines.pureharm.anomaly._
 import busymachines.pureharm.effects._
 import cats.syntax.all._
 
+/** backport integration of CE-3's implementation
+  * https://github.com/typelevel/cats-effect/blob/series/3.x/std/shared/src/main/scala/cats/effect/std/Random.scala#L342
+  *
+  * @param F
+  */
 abstract class RandomImpl[F[_]](implicit F: Sync[F]) extends Random[F] {
 
   protected def randomF: F[ScalaRandom]
 
   protected def delay[A](f: ScalaRandom => A): F[A] = randomF.flatMap(sr => F.delay(f(sr)))
 
-  override def boolean: F[Boolean] = delay(_.nextBoolean())
+  private[this] def require(condition: Boolean, errorMessage: => String): F[Unit] =
+    if (condition) ().pure[F]
+    else new IllegalArgumentException(errorMessage).raiseError[F, Unit]
 
-  override def int: F[Int] = delay(_.nextInt())
-  override def int(i: Int): F[Int] = delay(_.nextInt(i.max(1)))
+  override def betweenDouble(minInclusive: Double, maxExclusive: Double): F[Double] =
+    for {
+      _ <- require(minInclusive < maxExclusive, "Invalid bounds")
+      d <- nextDouble
+    } yield {
+      val next = d * (maxExclusive - minInclusive) + minInclusive
+      if (next < maxExclusive) next
+      else Math.nextAfter(maxExclusive, Double.NegativeInfinity)
+    }
 
-  override def int(l: Int, h: Int): F[Int] = for {
-    _ <-
-      if (l > h) InvalidInputAnomaly(s"Random.int(l: Int, h: Int) - l < h, but l=$l, h=$h").raiseError[F, Unit]
-      else F.unit
-    diff = h - l
-    lower <- int(diff)
-  } yield lower + diff
+  override def betweenFloat(minInclusive: Float, maxExclusive: Float): F[Float] =
+    for {
+      _ <- require(minInclusive < maxExclusive, "Invalid bounds")
+      f <- nextFloat
+    } yield {
+      val next = f * (maxExclusive - minInclusive) + minInclusive
+      if (next < maxExclusive) next
+      else Math.nextAfter(maxExclusive, Float.NegativeInfinity)
+    }
 
-  override def long: F[Long] = delay(_.nextLong())
+  override def betweenInt(minInclusive: Int, maxExclusive: Int): F[Int] =
+    require(minInclusive < maxExclusive, "Invalid bounds") *> {
+      val difference = maxExclusive - minInclusive
+      for {
+        out <-
+          if (difference >= 0) {
+            nextIntBounded(difference).map(_ + minInclusive)
+          }
+          else {
+            /* The interval size here is greater than Int.MaxValue,
+             * so the loop will exit with a probability of at least 1/2.
+             */
+            def loop(): F[Int] =
+              nextInt.flatMap { n =>
+                if (n >= minInclusive && n < maxExclusive) n.pure[F]
+                else loop()
+              }
+            loop()
+          }
+      } yield out
+    }
 
-  override def long(l: Long): F[Long] = delay(_.nextLong(l.max(1L)))
+  override def betweenLong(minInclusive: Long, maxExclusive: Long): F[Long] =
+    require(minInclusive < maxExclusive, "Invalid bounds") *> {
+      val difference = maxExclusive - minInclusive
+      for {
+        out <-
+          if (difference >= 0) {
+            nextLongBounded(difference).map(_ + minInclusive)
+          }
+          else {
+            /* The interval size here is greater than Long.MaxValue,
+             * so the loop will exit with a probability of at least 1/2.
+             */
+            def loop(): F[Long] =
+              nextLong.flatMap { n =>
+                if (n >= minInclusive && n < maxExclusive) n.pure[F]
+                else loop()
+              }
+            loop()
+          }
+      } yield out
+    }
 
-  override def long(l: Long, h: Long): F[Long] = for {
-    _ <-
-      if (l > h) InvalidInputAnomaly(s"Random.long(l: Int, h: Int) - l < h, but l=$l, h=$h").raiseError[F, Unit]
-      else F.unit
-    diff = h - l
-    lower <- long(diff)
-  } yield lower + diff
+  override def nextAlphaNumeric: F[Char] = {
+    val chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
+    nextIntBounded(chars.length()).map(chars.charAt(_))
+  }
 
-  override def double(h: Double): F[Double] = this.double(0.0d, h)
-  override def double(l: Double, h: Double): F[Double] = delay(_.between(l, h))
+  override def nextBoolean: F[Boolean] = delay(_.nextBoolean())
 
-  override def string(length: Int): F[String] =
+  override def nextBytes(n: Int): F[Array[Byte]] = delay { r =>
+    val bytes = new Array[Byte](0.max(n))
+    r.nextBytes(bytes)
+    bytes
+  }
+  override def nextDouble: F[Double] = delay(_.nextDouble())
+  override def nextFloat:    F[Float]  = delay(_.nextFloat())
+  override def nextGaussian: F[Double] = delay(_.nextGaussian())
+  override def nextInt:      F[Int]    = delay(_.nextInt())
+  override def nextIntBounded(n: Int): F[Int] = delay(_.nextInt(n))
+  override def nextLong: F[Long] = delay(_.nextLong())
+  override def nextLongBounded(n: Long): F[Long] = delay(_.nextLong(n))
+  override def nextPrintableChar: F[Char] = delay(_.nextPrintableChar())
+  override def nextString(length:  Int):       F[String]    = delay(_.nextString(length))
+  override def shuffleList[A](l:   List[A]):   F[List[A]]   = delay(_.shuffle(l))
+  override def shuffleVector[A](v: Vector[A]): F[Vector[A]] = delay(_.shuffle(v))
+
+  override def printableString(length: Int): F[String] =
     delay(sr => List.range(0, length).map(_ => sr.nextPrintableChar()).mkString)
 
   // see point 2 of the procedure: